app/requirements.txt
python-dotenv
pyserial
typer
